# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ariadne Daemon API",
    "description": "REST API of ariadned, the coding-agent orchestrator daemon.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.10.0"
  },
  "paths": {
...
    "/v1/tasks/{id}/transitions": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Transition audit log of a task.",
        "operationId": "list_transitions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "task id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskTransitionDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Request a status transition. The actor is derived from the call context.",
        "operationId": "transition",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "task id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentUsageDto": {
        "type": "object",
        "description": "What one staffed agent spent on a task, named the way a reader addresses\nit: an agent has no name of its own, so its skills are what identify it.",
        "required": [
          "agent_id",
          "skills",
          "usage"
        ],
        "properties": {
          "agent_id": {
            "type": "string"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The skills the agent loads; empty only if the agent is gone."
          },
          "step": {
            "type": [
              "string",
              "null"
            ]
          },
          "usage": {
            "$ref": "#/components/schemas/TokenUsageDto"
          }
        }
      },
...
      "Landing": {
        "type": "string",
        "description": "How the tasks of one goal end.\n\nThe one thing about the end of a task the author has to be told, since the\ncommands it runs differ entirely between the endings. The user chooses it\nonce, for the whole goal, when the goal is created, and every task of the\ngoal follows it: some work lands on the base branch, some goes through a\nrequest the author then sees to its merge, and some has nothing to land at\nall — a report filed, a document published, a release cut. Every ending\nreaches [`TaskStatus::Finished`]; landing is one way of getting there\nrather than the meaning of being there.\n\nWhich forge a published request goes to is *not* here: `origin` says\nwhether it is GitHub or GitLab, and asking the remote at landing time\ncannot go stale the way a second copy of the answer would.",
        "enum": [
          "merge",
          "pull_request",
          "none",
          "feature_branch"
        ]
      },
...
      "Seat": {
        "type": "string",
        "description": "Where an agent sits: the orchestrator of a goal, or the author or a\nreviewer of one task.\n\nA seat is a position, not an identity. Every agent below the orchestrator\nis generic, and what it can do comes from the skills it loads; the seat is\nonly what the state machine and the launcher need to know about it.",
        "enum": [
          "agent",
          "orchestrator",
          "author",
          "reviewer"
        ]
      },
...
      "TaskAgentDto": {
        "type": "object",
        "required": [
          "id",
          "seat",
          "skills",
          "model"
        ],
        "properties": {
          "branch": {
            "type": [
              "string",
              "null"
            ],
            "description": "The branch this agent works on: the task branch for the first author,\na suffixed sibling of it for every later one. None for a reviewer,\nwhich owns no branch."
          },
          "brief": {
            "type": [
              "string",
              "null"
            ],
            "description": "What the orchestrator told this agent beyond the task itself. None =\nthe task is the whole of it."
          },
          "effort": {
            "type": [
              "string",
              "null"
            ],
            "description": "The reasoning effort that model is run at. None = whatever the agent\nruns it at on its own.",
            "example": "high"
          },
          "id": {
            "type": "string"
          },
          "model": {
            "type": "string",
            "description": "What this agent runs on, `<agent>:<model>`.",
            "example": "codex-acp:o3"
          },
          "seat": {
            "$ref": "#/components/schemas/Seat",
            "description": "`author` or `reviewer`."
          },
          "session_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "This agent's live session, the id `POST /v1/sessions/{id}/switch`\ntakes. None while it carries no live session."
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The skills this agent loads, in the order they reach it.",
            "example": [
              "coding",
              "documentation"
            ]
          },
          "step": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "TaskDto": {
        "type": "object",
        "required": [
          "id",
          "goal_id",
          "repo_id",
          "title",
          "description",
          "status",
          "agents",
          "depends_on",
          "branch",
          "landing",
          "stalled",
          "picks",
          "usage",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "agents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskAgentDto"
            },
            "description": "The agents staffed on the task: the authors first, then the reviewers\nin review order. What each one can do is the skills it carries. Most\ntasks staff one author; one staffed with several runs them in\nparallel, and the reviewers pick the branch that lands."
          },
          "branch": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Ids of tasks that must merge before this one starts."
          },
          "description": {
            "type": "string"
          },
          "goal_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "landing": {
            "$ref": "#/components/schemas/Landing",
            "description": "How the task ends: a read-only copy of its goal's landing, which every\ntask of the goal follows."
          },
          "merge_commit": {
            "type": [
              "string",
              "null"
            ]
          },
          "picked_agent_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "The author the reviewers picked, on a task staffed with several: the\none whose branch lands. None for a one-author task, and until the\npick settles."
          },
          "picks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskPickDto"
            },
            "description": "The picks the reviewers have recorded so far, oldest first. Empty for\na one-author task."
          },
          "pr_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "URL of the pull or merge request the task was published as, once its\nauthor has reported one; None for a task landed directly."
          },
          "reason": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why a `failed` or `cancelled` task ended — the author's own\n`fail_task` reason, a dependency that never landed, a cancelled goal.\nNone for every other status, and for an ending nobody gave a reason\nfor."
          },
          "repo_id": {
            "type": "string",
            "description": "Id of the repository the task works in, one of its goal's."
          },
          "stalled": {
            "type": "boolean",
            "description": "Set when the agent went idle without advancing the task."
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "step": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/TaskUsageDto",
            "description": "What the agents of this task have spent between them."
          },
          "worktree_path": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "TaskPickDto": {
        "type": "object",
        "description": "One agent staffed on a task: where it sits, what it knows, and what it\nruns on.\n\nThe agent has no identity of its own. `seat` says only whether it authors\nthe task or reviews it; the skills are what it can do. What it runs on was\nsized by the orchestrator when it staffed the task, or chosen by the user\nsince — either way it is what this agent runs on, and nothing behind it\nchanges that.\nOne reviewer's pick of the winning author, on a task staffed with several\nauthors.",
        "required": [
          "reviewer_agent_id",
          "author_agent_id",
          "created_at"
        ],
        "properties": {
          "author_agent_id": {
            "type": "string",
            "description": "The author it picked."
          },
          "created_at": {
            "type": "string"
          },
          "reviewer_agent_id": {
            "type": "string",
            "description": "The reviewer that picked. One pick per reviewer per task."
          }
        }
      },
      "TaskStatus": {
        "type": "string",
        "description": "Task lifecycle status.",
        "enum": [
          "pending",
          "ready",
          "in_progress",
          "under_review",
          "changes_requested",
          "approved",
          "finished",
          "cancelled",
          "failed"
        ]
      },
      "TaskTransitionDto": {
        "type": "object",
        "required": [
          "id",
          "from_status",
          "to_status",
          "actor",
          "created_at"
        ],
        "properties": {
          "actor": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "from_status": {
            "type": "string"
          },
          "from_step": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "to_status": {
            "type": "string"
          },
          "to_step": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "TaskUsageDto": {
        "type": "object",
        "description": "What a task cost, by who spent it: its author, its reviewers one entry\neach, and the total of every session on the task.",
        "required": [
          "agents",
          "total",
          "author",
          "reviewers"
        ],
        "properties": {
          "agents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentUsageDto"
            }
          },
          "author": {
            "$ref": "#/components/schemas/TokenUsageDto",
            "description": "The author's own, across every run of it."
          },
          "reviewers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentUsageDto"
            },
            "description": "One entry per reviewer that has a session on the task, every review\nround of it summed, in review order. A reviewer whose session has yet\nto report anything is listed with zeros; one that has never been\nspawned is not listed at all."
          },
          "total": {
            "$ref": "#/components/schemas/TokenUsageDto",
            "description": "Every session on the task summed, whatever its seat."
          }
        }
      },
...
      "TokenUsageDto": {
        "type": "object",
        "description": "Tokens spent, as the agents' own transcripts report them.\n\nAlways present and always a number: nothing reported is zero, not null.",
        "required": [
          "input_tokens",
          "cached_input_tokens",
          "output_tokens"
        ],
        "properties": {
          "cached_input_tokens": {
            "type": "integer",
            "format": "int64",
            "description": "The subset of `input_tokens` served from the prompt cache, so never\nadded to it.",
            "minimum": 0
          },
          "input_tokens": {
            "type": "integer",
            "format": "int64",
            "description": "Prompt tokens, cache reads and cache writes included.",
            "minimum": 0
          },
          "output_tokens": {
            "type": "integer",
            "format": "int64",
            "description": "Completion tokens, thinking and reasoning included.",
            "minimum": 0
          }
        }
      },
      "TransitionRequest": {
        "type": "object",
        "required": [
          "to"
        ],
        "properties": {
          "merge_commit": {
            "type": [
              "string",
              "null"
            ],
            "description": "Required when `to` is `finished`, unless the task lands nothing."
          },
          "reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "to": {
            "$ref": "#/components/schemas/TaskStatus"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}